feat(Bitcoin Cash): Add price charts
