+++ app/src/main/java/piuk/blockchain/android/ui/dashboard/DashboardFragment.kt
@@ -13,11 +13,11 @@ import android.view.View
 import android.view.ViewGroup
 import kotlinx.android.synthetic.main.fragment_dashboard.*
 import piuk.blockchain.android.R
-import piuk.blockchain.android.data.currency.CryptoCurrencies
 import piuk.blockchain.android.injection.Injector
 import piuk.blockchain.android.ui.balance.BalanceFragment
 import piuk.blockchain.android.ui.base.BaseAuthActivity
 import piuk.blockchain.android.ui.base.BaseFragment
+import piuk.blockchain.android.ui.charts.ChartsActivity
 import piuk.blockchain.android.ui.dashboard.adapter.DashboardDelegateAdapter
 import piuk.blockchain.android.ui.home.MainActivity
 import piuk.blockchain.android.util.AndroidUtils
@@ -36,9 +36,7 @@ class DashboardFragment : BaseFragment<DashboardView, DashboardPresenter>(), Das
 
     @Inject lateinit var dashboardPresenter: DashboardPresenter
     private val dashboardAdapter by unsafeLazy {
-        DashboardDelegateAdapter(context!!) {
-            presenter.onAssetSelected(it)
-        }
+        DashboardDelegateAdapter(context!!) { ChartsActivity.start(context!!, it) }
     }
     private val receiver = object : BroadcastReceiver() {
         override fun onReceive(context: Context, intent: Intent) {
@@ -104,20 +102,8 @@ class DashboardFragment : BaseFragment<DashboardView, DashboardPresenter>(), Das
         dashboardAdapter.notifyItemChanged(position)
     }
 
-    override fun updateChartState(chartsState: ChartsState) {
-//        dashboardAdapter.updateChartState(chartsState)
-    }
-
     override fun updatePieChartState(chartsState: PieChartsState) {
         dashboardAdapter.updatePieChartState(chartsState)
-    }
-
-    override fun updateCryptoCurrencyPrice(price: String) {
-//        dashboardAdapter.updateCurrencyPrice(price)
-    }
-
-    override fun updateDashboardSelectedCurrency(cryptoCurrency: CryptoCurrencies) {
-//        dashboardAdapter.updateSelectedCurrency(cryptoCurrency)
     }
 
     override fun showToast(message: Int, toastType: String) {
