+++ app/src/main/java/piuk/blockchain/android/ui/dashboard/DashboardView.kt
@@ -1,7 +1,6 @@
 package piuk.blockchain.android.ui.dashboard
 
 import android.support.annotation.StringRes
-import piuk.blockchain.android.data.currency.CryptoCurrencies
 import piuk.blockchain.android.ui.base.View
 import piuk.blockchain.android.ui.customviews.ToastCustom
 import java.util.*
@@ -12,11 +11,7 @@ interface DashboardView : View {
 
     val locale: Locale
 
-    fun updateChartState(chartsState: ChartsState)
-
     fun updatePieChartState(chartsState: PieChartsState)
-
-    fun updateCryptoCurrencyPrice(price: String)
 
     fun showToast(@StringRes message: Int, @ToastCustom.ToastType toastType: String)
 
@@ -29,7 +24,5 @@ interface DashboardView : View {
     fun startBuyActivity()
 
     fun startShapeShiftActivity()
-
-    fun updateDashboardSelectedCurrency(cryptoCurrency: CryptoCurrencies)
 
 }
