+++ app/src/main/res/layout/activity_graphs.xml
@@ -188,7 +188,7 @@
 
     <android.support.constraint.Guideline
         android:id="@+id/guideline"
-        android:layout_width="wrap_content"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:orientation="horizontal"
         app:layout_constraintGuide_begin="280dp" />
